GH-369: [refactor] Migrated posts events
